Day 9/bank_reconciliation/requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pandas>=2.1.3
streamlit>=1.28.2
python-dotenv>=1.0.0
requests>=2.31.0
chromadb>=0.4.22
langchain-community>=0.0.38
beautifulsoup4==4.12.3
faiss-cpu==1.7.4
rapidfuzz==3.9.1 